kloud: improve "team.whoami" method
